# docs/swagger.yaml
definitions:
...
  models.OrderRequest:
    properties:
      amount:
        type: number
      product:
        properties:
          id:
            type: string
          name:
            type: string
          quantity:
            type: integer
        type: object
    type: object
info:
  contact: {}
paths:
...
  /api/v1/orders/create-order:
    post:
      consumes:
      - application/json
      description: Creates a new order and returns the status
      parameters:
      - description: Order payload
        in: body
        name: order
        required: true
        schema:
          $ref: '#/definitions/models.OrderRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Create a new order
      tags:
      - orders
  /api/v1/orders/replay-failed-events:
    post:
      description: Replays failed order events that have not been successfully published
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Replay failed order events
      tags:
      - orders
swagger: "2.0"